third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/cascade/same-origin-with-own-policy-window-open-expected.txt

CONSOLE ERROR: Refused to load the image 'http://example.test:8000/resources/square.png?blank-frame' because it violates the following Content Security Policy directive: "img-src http://127.0.0.1:8000".

CONSOLE ERROR: Refused to load the image 'http://example.test:8000/resources/square.png?blob-frame' because it violates the following Content Security Policy directive: "img-src http://127.0.0.1:8000".

CONSOLE ERROR: Refused to load the image 'http://example.test:8000/resources/square.png?filesystem-frame' because it violates the following Content Security Policy directive: "img-src http://127.0.0.1:8000".

This is a testharness.js-based test.
PASS Policy applied in top-level. 
PASS Image loaded via about:blank window blocked. 
PASS Image loaded via 'blob:' window blocked. 
PASS Image loaded via 'filesystem:' window blocked. 
Harness: the test ran to completion.
